refactor(suite-native): share wrap and unwrap transaction machinery
